fin de journee 09/10

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: asaux <asaux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 15:45:08 by asaux             #+#    #+#             */
-/*   Updated: 2023/10/04 16:12:05 by asaux            ###   ########.fr       */
+/*   Updated: 2023/10/09 10:52:45 by asaux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n)
 	i = 0;
 	if (!src && !dest)
 		return (0);
-	while (i < n)
+	while (i < (int) n)
 	{
 		((char *) dest)[i] = ((char *) src)[i];
 		i++;
