+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: asaux <asaux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 15:17:12 by asaux             #+#    #+#             */
-/*   Updated: 2023/10/08 17:47:48 by asaux            ###   ########.fr       */
+/*   Updated: 2023/10/09 10:54:50 by asaux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,14 +23,14 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 	s = (char *)src;
 	if (d < s)
 	{
-		while (n-- >= 0)
+		while ((int) n-- >= 0)
 			*d++ = *s++;
 	}
 	else
 	{
 		d = d + n - 1;
 		s = s + n - 1;
-		while (n-- >= 0)
+		while ((int) n-- >= 0)
 			*d-- = *s--;
 	}
 	return (dest);
