+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: asaux <asaux@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 11:32:27 by asaux             #+#    #+#             */
-/*   Updated: 2023/10/08 17:54:23 by asaux            ###   ########.fr       */
+/*   Updated: 2023/10/09 10:59:54 by asaux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 	i = 0;
 	if (size > 0)
 	{
-		while (sc[i] && i < size - 1)
+		while (sc[i] && i < (int) size - 1)
 		{
 			dst[i] = sc[i];
 			i++;
